// react-frontend/src/api/models/auth-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AuthResponse
 */
export interface AuthResponse {
    /**
     * 
     * @type {string}
     * @memberof AuthResponse
     */
    'accessToken'?: string;
    /**
     * 
     * @type {string}
     * @memberof AuthResponse
     */
    'refreshToken'?: string;
    /**
     * 
     * @type {string}
     * @memberof AuthResponse
     */
    'tokenType'?: string;
    /**
     * 
     * @type {number}
     * @memberof AuthResponse
     */
    'expiresIn'?: number;
    /**
     * 
     * @type {string}
     * @memberof AuthResponse
     */
    'firstName'?: string;
    /**
     * 
     * @type {string}
     * @memberof AuthResponse
     */
    'lastName'?: string;
    /**
     * 
     * @type {string}
     * @memberof AuthResponse
     */
    'email'?: string;
    /**
     * 
     * @type {Set<string>}
     * @memberof AuthResponse
     */
    'roles'?: Set<string>;
}